Initial layout support for SIXEL graphics.
When SIXEL data is encountered, the terminal now finds
the apparent number of text columns and rows that would
be covered by the image and offsets the text cursor to
make room.  Each affected cell is then assigned a bitmap
ID.  The bitmap ID scheme has changed slightly to ensure
that the “entire” NSImage is always available from any
cell, and each cell renders a segment of the full image.
(This will also be used later for things like contextual
menu items that operate on an image regardless of which
terminal cell is clicked by the user.)

Terminals are not yet displaying SIXEL-derived images
but the image display mechanism does work, verified by
temporarily constructing a raw test image and having
adjacent terminal cells each render a portion of it.

Note that since the decoder is still in debug mode, any
particularly large image is likely to trigger a lot of
console messages and the response may be slow.  These
messages will eventually be turned off by default and
controlled by a checkbox in the Debug Interface (much
like other terminal state machine output).