Support for iTerm2 base64 image sequences.
A new sub-emulator has been added to recognize the custom
terminal sequences defined by iTerm2, initially handling
images that are sent as base64-encoded strings (since work
was recently done to support Sixels, displaying images in
the terminal is now relatively easy).  The majority of
iTerm2 image parameters are also supported, such as a
suggested number of terminal cells to occupy.

The Preferences window has been extended to have a tweak
for enabling or disabling iTerm graphics support.  Also,
if a file name is included in the data stream, it now will
briefly appear as a floating tooltip over the cursor as
the image is displayed.

An extension has been added to NSData that just makes it
easier to access base64 support that was already in the OS
and not declared conveniently by the current SDK.