Images may now be saved; updated Help.
The File >> "Save Selection..." command now creates an
image file (in any supported format, based on extension)
if the highlighted region crosses a bitmap image cell.
The original image is always used for saving, regardless
of how much of the image is selected or otherwise shown
by the terminal.  Also, depending on the source of the
data, the saved image may actually look better than what
is displayed by the terminal itself.

Internally, the mechanism for looking at ranges of text
attributes has been converted to use a block instead of
a callback function.  This is now used not only to draw
but also to quickly locate images in selections.  In the
near future, more commands are likely to be supported
besides saving (e.g. Copy is a likely candidate).

Help pages have been updated to mention that images can
now be selected, and to describe recent bitmap image
additions in more detail.