Changed device control strings to be more general.
The DCS / ESC-P sequence (used by Sixel data but also by
other sequences in some terminals) is now handled in a
more general way.  The entire string up to the terminator
is captured by one of the base VT emulators prior to any
processing, including parameter parsing, which means that
the main state machine no longer has to try to handle
parameters inside a device control string.  Also, it is
no longer necessary for the Sixel reader or any other
user of a device control string to maintain extra states
for capturing the string properly.

The new parameter-decoding object (used initially only
for Sixel data within a device control string) has been
improved to track termination as a separate state and
use the final parameter as its accumulator.  This
prevents certain inputs from allocating a number of
useless undefined parameters at the end of the list.
The unit test functions have been updated to show the
changes in behavior as well.