Terminal debugging log is now more compact.
Although logging is only done when it is enabled by the
debugging interface, it previously had a rather verbose
format that made states difficult to keep track of.
Now, layers of callbacks (such as XTerm versus VT100)
are tracked in sequence and the “path” to a state is
logged on a single line instead of perhaps many lines.