Secondary device attributes use XTerm patch level.
A real XTerm will hijack the VT220 “secondary device
attributes” query to return the XFree86 patch level
in place of the “firmware version” field.  Since some
terminal applications actually expect this behavior,
MacTerm now does the same, for XTerm terminals only.
The minimum value of 95 is now chosen, instead of the
previous value of 10 (a.k.a. “firmware version 1.0”).
If absolutely necessary, this value can now be reset
using a new low-level preference, as documented in
the release notes and MacTerm Help.