Initial support for DECRQM, DECRQSS and DECSCUSR.
More changes to reduce issues with recent versions of
"vim": this adds initial support for more terminal
sequences, namely to respond to certain queries and
to accept cursor shape-change requests.  Currently,
cursor changes are stored only for reporting purposes
and they do not propagate to the display.  Note that
there are dozens of additional queries that ought to
be supported, along with corresponding sequences to
set the same values; those enhancements will be
considered for later commits.