Fixed window rotation commands.
Sometimes a command like “Bring Next to Front” would do nothing,
and this turned out to be a fundamental bug in the strict weak
ordering function used to map windows by screen (entire sets of
windows might just be skipped, based on values in the map).  This
has been fixed, and the internal API for switching windows has
been cleaned up a bit too.  A special check has also been added
for Carbon drawers, as a short-term way to prevent terminal tab
titles from being included in the window rotation.