Initial setup for arbitrary new Cocoa windows.
This is meant to improve integration of the experimental
Cocoa terminal windows with the rest of the application.
The Terminal Window object can now be told to use only
Cocoa (no Carbon), and the Debug Interface can be used
to turn this default on or off for any new terminals.

In this first version, the windows are not that useful;
they show a title, and actions in the Session Info list
(like renaming or double-clicking) will work on them.
A variety of errors are logged, and these will be fixed
incrementally.  Cocoa terminal views will not appear in
these windows yet; the new views are created but there
is no sizing or layout code in place and no code to
process events.

Note that, as before, the Debug Interface also exposes
a singleton Cocoa terminal window with more features
such as real views and the ability to bind active
sessions.  (The legacy Carbon window is needed for any
keyboard or mouse input.)  This will eventually be
removed when it is possible to do the same kinds of
tests with the refactored Terminal Window object.