New menu commands for tabs in Cocoa terminals.
Updated main menu commands to include the new system
actions that work with window tabs.  Items have been
shuffled and renamed a bit to ensure that actions do
not appear twice (as the OS has a mechanism for
automatically inserting items if it cannot find the
custom items).  Standard actions do not have exactly
the same names that Apple specifies, in part to be
more consistent with previous MacTerm commands and
existing MacTerm concepts such as Workspaces; for
example, “Move to New Workspace” is still the name
of the command for moving a tab to a new window and
not “Move Tab to New Window”, even though the item’s
selector has changed to match the standard one.

The new Cocoa terminal windows (still only usable
via the debug interface) therefore now support more
tab management features, such as the ability to see
all tabs in an Exposé-style view and the ability to
combine multiple windows into one set of tabs.

Where possible, legacy Carbon windows that are using
the old-style tab drawers will provide the same tab
management commands as before, except using the new
action selectors.  Any other actions, exclusive to
Cocoa (such as the Exposé-style view of all tabs),
will be disabled in Carbon windows.

Session management has not yet been fixed to make
Cocoa windows automatically turn into tabs, such as
when a tabbed Workspace is opened.