Add responder-chain bias to Commands module.
New utility methods consolidate some of the logic that
allowed actions to check for Carbon windows and handle
menu commands, etc. differently for Carbon targets.
Now, every single action that is implemented by the
Commands_Executor object will now allow the first
responder in Cocoa to implement the action, provided
the main window is not a Carbon window.  This allows
Cocoa windows (such as the new terminal window) to
gradually implement commands in a more normal fashion
without breaking the Carbon work-arounds.

Also, this fixes a few minor command issues such as
the “Show/Hide Toolbar” command not working for Cocoa
terminal windows.