Improve robustness of key-value observing.
All implementations of key-value observers have been
fixed to pass unrecognized cases to the superclass.
This required unique “context” values so the helper
methods in the Cocoa Extensions module have been
updated to automatically set a unique context (using
the observer-tracking object that is already created).
A new "observerArray:containsContext:" extension was
added to make it really easy for objects to tell when
a given context corresponds to one of that object’s
observers.