Add Close/Minimize/Zoom to customizable toolbars.
Experimental Cocoa terminal windows now provide toolbar
items for the standard window buttons (Close, Minimize
and Zoom, where the latter replaces the Full Screen item)
and the normal top-left cluster of 3 buttons is hidden.
This gives the user a lot of power, including the ability
to remove some or all of the window buttons or relocate
them to different positions in the toolbar!  Also, since
toolbar items remain visible in Full Screen mode, it is
not necessary to use the mouse to show the menu bar and
wait to click the window buttons while in Full Screen;
you can simply click on them immediately.

If “Hide Toolbar” is used, the normal top-left cluster
of 3 window buttons reappears.  Customizations to window
elements do not apply when the toolbar is hidden.