Prevent initial sound in Notifications tab.
Preferences window General pane Notifications tab no
longer plays a sound when it is first created (as a
side effect of binding the bell-sound preference).
The sound is only played when the user makes changes.
This required creating a new property to keep track
of the “loaded” state of the view.