Fix enabled states of Preferences list actions.
In the Preferences window while a list of collections
is displayed, certain actions can now be disabled
based on the selected item.  (Previously the actions
would trigger an error sound if they were chosen by
the user for an invalid item.)  Currently, the only
item that may not support all actions is the Default
item.  When Default is selected, the deletion (minus)
button is dimmed and the commands “Rename Collection”
and “Copy Collection Settings to Default…” are also
unavailable.