Add new default env. var. for Cocoa terminals.
New MACTERM_USE_COCOA_TERM environment variable can be
set to "1" (e.g. when launching MacTerm from a shell)
to automatically use the experimental new Cocoa windows
for all new sessions.  It is still possible to change
this from the Debug Interface but the env. variable
turns this on by default.  The Debug Interface has also
been simplified, removing the previous options to spawn
a “test” Cocoa terminal or bind that test terminal to a
session, since it is now possible to spawn arbitrary
Cocoa terminal windows.