Create delegating-layout view.
Since a view’s controller is likely to maintain complex
child view information (such as an array of other view
controllers), it would be convenient to have all of that
available for subview layout purposes.  Unfortunately,
NSView only implicitly notifies the view itself if the
frame changes; the controller is NOT told.  It is not
good to require special view subclasses to track views
for layout purposes when their controllers are already
tracking those same views.

The problem is addressed in this commit by adding a new
protocol and a new default NSView subclass.  This view
calls its “layout delegate” whenever the view’s frame
changes.  Also, a new method has been added to NSView so
that layout can be performed upon request.

Note that this is being done for simplicity and to allow
older SDKs.  For example, auto-layout is not available
(and might have efficiency and complexity issues anyway).