Remove useless gap on left side of Cocoa toolbar.
This removes the gap on the left side of the toolbar in
Cocoa terminal windows (that was reserved by the system
for standard window buttons that are hidden).  Standard
window buttons are hidden because the terminal toolbar
provides items for Close, Minimize and Zoom that allow
the user to decide if and where those buttons should
take up valuable window space.  Also, the toolbar items
remain visible in Full Screen.

Note that this required overriding a private method of
NSWindow so the layout may break again in the future.