Initial Cocoa terminal text input implemented.
Added initial “wiring” for translating key events from
the new Cocoa terminal views to the open session.  This
works for basic text strings, tabs, newlines and delete
keys.  It does not yet work for control keys or other
cases that may require special interpretation of the
original key event.  As a work-around for now, other
input methods can be used such as the floating keypads
or the floating Command Line window.

This is implemented with a new protocol and delegate
mechanism, whereby the Terminal View does not need to
know that it is sending events to a Session.  (Rather,
the Session allocates an input handler internally that
receives Objective-C calls as needed and performs the
appropriate actions such as sending text or special
sequences to the session process.)  Also, even though
the Terminal View must handle the entire Cocoa standard
text input protocol, unnecessary complexity is removed
when communicating with the delegate: the Session only
has to respond to terminal-relevant input cases.