Major enhancements to vector graphics windows.
New Cocoa event-handling has been added to the vector
graphics window, restoring some mouse functionality that
existed many years before to zoom in on part of an image.
The zoom has been modernized in appearance, and also in
behavior (by supporting trackpads for instance).  Also,
the zoom is no longer implemented in the vector graphics
interpreter; graphical transforms are applied to the
canvas view instead.

Like terminal views, vector graphics windows now have a
“matte” and they use the same colors from any assigned
Format.  In addition, the matte background color is used
for the extended background if the user zooms out images
really far.  The Vector Window module has been extended to
handle the matte, although the Vector Canvas is able to
assign the correct color to the matte just by scanning its
superview hierarchy.

The Cocoa Terminal Background view that is used for mattes
can now capture and forward clicks, as its Carbon version
did.  This allows the user to be more “sloppy” about any
clicks near the edge of content: in the case of vector
graphics, it means that the user can click the matte and
this will be translated into a mouse-down at the nearest
valid point in the graphic to start a rectangle drag.
(This enhancement will also be needed later for the mattes
of Cocoa-based terminal windows.)

Finally, adding Cocoa code for events and drawing allows
some older Carbon-only code to be removed, as the vector
canvas was the last thing depending on the old code.