Version 4.1 now FINALIZED.  5.0 alpha to begin.
After additional testing and years of development, this is a
reasonable point to stabilize the 4.1 series.  Release notes
and version data have been updated to 4.1.0 final.

In the near future, development work on Cocoa, etc. will become
the default for a 5.0 alpha series, the SDK will be updated and
the minimum required OS version may increase.