Initial 5.0 alpha release.
This updates the application version, adds warning
information to the About box, and switches the default
terminal implementation from Carbon to Cocoa (where
now the preference in the Debug Interface is to revert
to Carbon, instead of a preference to enable Cocoa).
Note that this change will introduce considerable bugs
because there are known problems with the Cocoa window
(everything from text rendering to certain features
not working correctly).