Add Cocoa terminal receive-drop; update Clipboard.
Removed drag-and-drop code from legacy Carbon view, and
added recognition of URL and text drops to the Cocoa
terminal view.  Also reimplemented Clipboard functions
in terms of NSPasteboard*, removing dependencies on the
older types and other unused functions.  And, updated
the Clipboard window to use NSPasteboard*.  Overall, the
amount of code required to implement these features is
significantly reduced.