Remove “Cocoa Future” module.
The various Objective-C hacks used to access conditional
APIs in the past are no longer necessary because of a
combination of the latest SDK and new compiler features
such as "@available" syntax.  Previous dependencies on
the Cocoa Future module have been updated to use either
direct references or "@available" blocks, and the entire
Cocoa Future module has been removed.