Initial image asset infrastructure.
Converted Full Screen “off switch” icon into an image set
in an assets folder so that it can automatically use a
different color in Dark Mode.  (This is a simple test to
serve as a foundation for gradually migrating all the
other icons in a similar way.)  Updated the build to have
a separate ApplicationMedia.xcassets source that merges
automatically with SharedMedia.xcassets when producing
the main bundle’s "Assets.car" file.