Migrate most icon artwork to template images.
Most icons, presented primarily in toolbars, are now
images stored in an asset catalog (“2x” only, so far).
As “template” images, the original icons automatically
appear mostly white when the system is in Dark Mode.
Most icons could be migrated as-is but a few required
small changes to the artwork source files (which were
then exported as PNG).