Changes to appease "hiutil" in macOS Mojave.
The "hiutil" command in macOS Mojave cannot open generated
".helpindex" files and the error message is very unhelpful.
The same error message is printed even when trying to read
index files for Apple apps such as Xcode!  Therefore, the
errors from this post-validation step are now suppressed
to prevent the first build from failing.

Also, "hiutil" seemingly refuses to read files that do not
have "robots" meta information so this is now set in every
file.

The "hiutil" command was also spewing warnings about files
not containing anchors, and this appears to be its wrong
belief that the "<a>" tag is required (in reality, an "id"
from any tag is a valid anchor target).  Added dummy "<a>"
tags in all files to appease "hiutil".