Update commands for Session Info and speech.
Legacy command IDs for showing and hiding Session Info
have been removed, along with some APIs in the Commands
module that are no longer needed.  Speech commands have
also been migrated to selectors with names matching the
speech features of NSTextView, and they are now in the
top-level Terminal menu (not just in contextual menus).
Fixed a few cases of contextual menu items with targets
set to the Commands module instead of first responder.