Modernize Preferences window drag-and-drop.
Removed implementation of deprecated method for table
row drag-and-drop, and replaced with newer version for
the same purpose.  (Technically the newer alternative
had already existed from the beginning but it was never
used by the system because the method name was not
quite correct.  The latest compiler and SDK issued a
warning so this was finally addressed.)