Port graphics-glyph renderer to Cocoa.
Line-drawing and other special characters handled
by custom code are now visible in the Cocoa view.
The code has been simplified somewhat because a
number of text-like cases that previously had to
be handled separately, like Greek letters, are now
passed directly to Core Text.  This is not quite
complete, as some rendering artifacts are visible
(lines not touching adjacent cells, etc.).