Update default preferences for new installations.
Preferences for new users are now slightly different
(existing users will not see their settings change).
The default terminal type is now XTerm instead of
VT100, since the XTerm implementation is the most
powerful; this improves out-of-the-box support, as
modern programs tend to expect XTerm features.  Also,
MacTerm now opens a log-in shell by default instead
of the Custom New Session sheet; the previous choice
was more useful in the days of remote connections,
and now people simply expect a shell to be opened.
Finally, in Full Screen, the “off switch” is not on
by default anymore because the system provides its
own method for exiting Full Screen that people are
likely to be aware of.