Remove last direct dependency on Apple Events.
The “import” mechanism in the Preferences window was
still implemented by sending an open event back to
the application; this now directly calls the Quills
interface instead (to use Python bindings).  As this
was the last direct use of Apple Events APIs, several
source files related to AE have been removed.