Session can now have Dark/Light associated Format.
The “associated Format” feature of Session Preferences can
now be used to set different default Format settings for
Light and Dark modes of the OS.  If the system is currently
in Dark Mode when the session starts, the “dark” Format is
chosen for colors, etc.; otherwise, the “light” Format is
used.  Code has been added to detect changes to the system
Dark setting but this is not yet used to auto-switch the
Format of open terminal windows (this may be added later).