Replace tab titles with segmented controls.
Tab titles are now implemented as a separate “segmented
control” so tabs do not have a special background color.
This addresses display issues with sheets in Dark Mode
and it is a slightly better use of space.  Fortunately
most of the code for managing the tabs themselves is
unchanged because a tab view can function without any
visible titles.  There is a bit of new code to ensure
that the separate segments are updated when the tab view
changes and vice-versa.