Update background notification code.
Support for the Growl framework has been removed
entirely, as it is no longer feasible to test and
its main function has been replaced by not one but
two different Apple-provided system frameworks in
recent years.

The code for notifications has been simplified in a
number of ways, as the base SDK is much newer.  Some
of the annoyance has been toned down too, removing
notifications about sessions that exit successfully
for example.

Although the foundation has been laid to migrate to
the newest “user notification framework”, this does
not yet work due to runtime permission issues (most
likely related to ever-expanding GateKeeper features
in some way; of course, the errors are far too vague
and the documentation far too pathetic to be sure).
It will probably be fixed at some point in the future
but at a low priority, since NSUserNotificationCenter
still works.

This submission also includes a minor fix to restore
“background notification on bell” if that preference
is set by the user.