Add support for "__weak" references in build.
Now that the project uses a more recent SDK, it is
possible to tag references as "__weak" even though
ARC is not enabled throughout.  The build has been
updated to allow "__weak" variables, and a few
known cases have been updated.  Note that a full
review has not been conducted and other pointers
may need to be identified as "__weak" later.