Migrate remaining commands to responder chain.
This finishes the work that started in several
recent commits to fully adopt the responder chain.
Also, any existing command selectors that were not
part of protocols are now mostly grouped that way.
Protocols provide a single place to declare the
right name of each command selector and state the
required/optional implementation of a group; and,
they clearly show where commands are implemented.

The only remaining use for command IDs is to handle
display commands (such as opening the Preferences
window to a particular pane).  Once these are
replaced, the Commands module can be simplified.