Restore scroll bar, and tick marks for “Find”.
A basic scroll bar knob is now displayed when there is a
scrollback buffer (this needs more work to be complete,
e.g. it does not respond to mouse drags).

New Cocoa-based scroll bars now display red tick marks
at the approximate locations of search results when
“Find” is used.  This feature previously existed for the
older Carbon-based view but it was not ported.