Add session icons to “Workspaces” panel.
In the Preferences window, “Workspaces” panel, “Windows” tab,
the list now shows an icon for each window that has an assigned
Session.  This helps to see how many windows are actually going
to be spawned by the Workspace, and the commands they will run.

A similar icon column has been added to “Macros” too but it is
currently not changed per macro.  (A future update might use
this to indicate at a glance the type of action assigned to a
macro in the list; or, perhaps a user-assigned icon.)