Make dynamic menu items more robust.
Certain menus generate dynamic lists of items
based on preferences, and in some cases the
menu was displayed by the system without these
extra items.  (The issue seems to be that the
menu object pointer itself changes, without the
system notifying that this has happened.)  To
make this more robust, a new notification is
registered so that the menus are rebuilt just
before menu tracking begins.