Add terminal-associated macro sets.
New “Default Macro Set” preference on a session
allows specific macros to be active whenever that
session first starts.  In addition, any time the
current macro set is changed, that choice is
associated with the active terminal window and
the current macro set is restored when switching
back to the same window.