Migrate “Debug Interface” to SwiftUI.
(This also updates the minimum SDK to 10.15.)
This adds a top-level workspace and “playground”.
The framework build was updated slightly so that
Swift and the playground can easily refer to UI
code from the main framework.

To test new UI, just one line of code in the
playground is changed to choose a new interface.
This in many ways lays a nice foundation for all
future SwiftUI work, including some common code
for certain UI element settings and layouts.

As a first test of SwiftUI, the “Debug Interface”
panel has been converted to SwiftUI and its XIB
has been deleted.  Its bindings have been changed
to rely on a Swift data model object, and the
only direct tie to AppKit that remains is the
wrapping window controller.