Change alerts to resemble Big Sur style.
Since “standard” alert APIs are not used in MacTerm,
the UI updates in Big Sur do not automatically apply.
The layout of custom alerts has now been changed to
be more like those in macOS Big Sur.  (So far, the
frame style is unchanged.)

Of course, an advantage of having a custom UI is that
any issues with the standard UI can be avoided.  One
of those issues is text alignment; to make long text
more readable, MacTerm does not center-align help text
in alerts.  Also, custom alerts continue to support a
large title, separate from message text and help text.