Migrate “screen size” panel to SwiftUI.
This begins the process of supporting SwiftUI in
panels that are bound to preferences (which is most
of the UI right now).  Using “screen size” settings
as a starting point, this adds some new common code
that will make it easier to port other panels in
the near future.

Unlike most simple bindings, Preferences need to be
able to handle inheritance from Default.  This adds
some common UI code to make it easier for SwiftUI
interfaces to describe their Default relationships.

Also updated the base Panel view controller to allow
a view-based initializer instead of just a NIB-based
one (since SwiftUI does not use NIBs).  Updated the
dependent AlertMessages module accordingly.