Add missing tooltips to all SwiftUI views.
As of macOS Big Sur, it is now possible to set
tooltips via a help() method in SwiftUI.  It is
not possible to simply call ".help()" though,
if compatibility with older OS versions is
required!  As a workaround, the View extension
"macTermToolTipText()" was added so that code
can call it unconditionally on any OS version.
Within this, "if #available" is used to call
help() only when it exists.  And, since a view
builder always has to be able to return some
View, "offset(x: 0, y: 0)" was used as a no-op
in the chain for the case where help() is not
defined.