Fix NSMenuToolbarItem crash on macOS 10.15.
Apparently NSMenuToolbarItem does not implement the
method "isBordered" on macOS 10.15 but the OS calls
it anyway (in Preferences-styled toolbars), causing
a runtime exception in 10.15 (not in 11.0).  This
is now avoided by subclassing NSMenuToolbarItem
with an implementation of the "bordered" property.

Also tweaked icon sizes on 10.15 to make the toolbar
look better with the smaller-style menu toolbar items
in 10.15.