Improve various sheet UI features on Big Sur.
Sheets on Big Sur are now displayed using its
new animation (darkened background window).
Note that this required tweaking the style of
the “window rename” sheet, because Big Sur
completely overrides the frame appearance of
windows that are presented as sheets (in the
case of window rename, the frame was different
enough that the override was a display issue;
for other sheets, Big Sur merely affects the
frame color).

Tab-panel displays such as the Format sheet
now use the new “large size” segmented control
on Big Sur.

Also, fixed a clipping problem in sheets that
used both SwiftUI and the “Use Default” text.