Restore inlining for debug-interface queries.
Debug mode checks are once again implemented as
inline functions that read exposed variables.
(These had changed to use Objective-C properties
on a model object without inlining, for SwiftUI;
this proved to be inefficient, especially since
some queries are extremely frequent.)

SwiftUI is still used for the “Debug Interface”
to bind to a model object but a new callback
has been added to respond to changed settings.
This allows non-Objective-C variables to be in
sync, and used for inline querying functions.