Migrate “Keyboard” panel to SwiftUI; minor fixes.
This updates the panel used for specifying key
bindings (via “Custom Key Sequences…” or the
“Keyboard” tab in the Sessions preference pane).

Since there is now some overlap between the needs
of the “Control Keys” palette and the buttons for
editing key bindings (interrupt/suspend/resume),
the code for the keypad view was also updated.
Key binding buttons now look like keys, and have
help messages to describe what the buttons do.

Panels continue to require quite a bit of mapping
code, which ought to be avoided if enumerations
are described as Objective-C-bound-to-Swift and
not directly in SwiftUI.  More is needed here.

Minor corrections were made to mark comments for
protocols in some preference panels.