Fix certain warnings during the build.
Changed the form of scanFloat() in the Swift code
of PrintPreview, since the previous form is now
deprecated.

Updated dimensions of certain cursor images to
address warnings during ".tiff" construction in
the build.  Now, “small” cursor images at 1x size
have a height of 32 pixels (not 36 pixels), which
is exactly half the height of the 2x versions.
This also fixes scaling artifacts in the cursors.