Fix menu commands for vector graphics windows.
Format menu items are once again supported for
vector graphics windows, allowing the colors of
the foreground and background to be changed.
The Format’s mouse cursor color is also used.

Added a new “Actual Size” command for graphics
windows, as a more explicit way to trigger the
return-to-normal-zoom (previously only possible
by double-clicking the canvas).

Added contextual menus to graphics windows.