Remove enum-mapping code for various UI settings.
Certain enumerations that directly map to SwiftUI
values have moved to the framework header and have
been given Swift equivalents using CF_...ENUM().
This allows SwiftUI to directly refer to the values
instead of having redundant UI-only enumerations,
and code to map to/from values is no longer needed.
This allows about 500 lines of code to be deleted!