Restore terminal resize (font size or dimensions).
Fixed new Cocoa-based terminal window implementation
to support live resize of the terminal view itself,
which was a feature of the previous implementation.
As before, the Control key can be used to temporarily
toggle the preference setting (which determines what
changes — font size or screen dimensions — as the
window is resized).

Slightly modified the “bubble” that is used to show
live resize information, to have more flexibility;
added a method to explicitly reset the shared window
(requiring minor updates to all the code using it).
Also, the format strings used to show live resize
properties are now exported for localization.