Migrate preference panel code to ARC.
ARC is now used when compiling code for the
Preferences window and all preference panels.
Generally, this reduces the complexity, as
several "dealloc" methods went away and a lot
of array expressions are just simpler now.

The signature of "context:" parameters in the
base Panel API has changed from "void*" to
"NSObject*" to make it straightforward for ARC
to understand reference transfers.  (Existing
uses of this parameter were all Objective-C
objects underneath anyway.)