Fix possible crash when decoding giant images.
Large images are now gracefully truncated to fit
the number of columns and/or rows of a terminal
screen, instead of crashing.  (For example, this
could happen when "imgcat" Base64-decoded a very
large image that would not fit in the window.)

Previously truncation was only detected in debug
mode with certain logging options enabled, and
cell dimensions were not actually modified to
deal with the error.  Now, the cell range for an
image is always forced to be valid.

Updated Base64 API to fix deprecation warning.