Migrate vector graphics view/window to ARC.
Enabled automatic reference counting for code that
deals with vector graphics (canvas view and window).
Also, modernized the headers (nullable annotations,
etc.), removing data members from the headers and
relocating documentation to property declarations.

To make the vector graphics window work better with
ARC, it now has a small internal data structure
similar to what is done for terminal windows.  The
structure holds a strong reference to the window
controller, which previously was the type that was
cast to the external reference.  Now, the reference
is cast to the internal structure pointer.