Migrate Rename interface to ARC.
Enabled automatic reference counting for Rename.
Also, modernized the headers (nullable annotations,
etc.), removing data members from the headers and
relocating documentation to property declarations.

The WindowTitleDialog_Ref type has been redefined
as a pointer to an Objective-C object so that it
works seamlessly with ARC.

Fixed a case where previously discarding this UI
might cause that title value to still be displayed
the next time the UI is used.