Migrate Server Browser interface to ARC.
Enabled automatic ref. counting for Server Browser.
Also, modernized the headers (nullable annotations,
etc.), removing data members from the headers and
relocating documentation to property declarations.

The primary ServerBrowser_Ref type has been redefined
as a pointer to an Objective-C object so that it
works seamlessly with ARC.  This makes the API a bit
smaller too.

Updated corresponding code in PrefPanelSessions to
use the modified browser type, and cleaned up that
module’s properties, etc. in a similar way.