Migrate PopoverManager interface to ARC.
Enabled automatic ref. counting for PopoverManager.
Also, modernized the headers (nullable annotations,
etc.), removing data members from the headers and
relocating documentation to property declarations.

The primary PopoverManager_Ref type has been redefined
as a pointer to an Objective-C object so that it
works seamlessly with ARC.  This makes the API a bit
smaller too.  Existing properties that stored the
reference as a simple "assign" value are now "strong".