Migrate ChildProcessWC module to ARC; minor fixes.
Enabled automatic ref. counting for ChildProcessWC.
Also, modernized the header (nullable annotations,
etc.).

Minor enhancements to the “child process window”
mechanism, including a custom window class.  This
is used to detect the case where the application
may seem to have visible open windows but all of
the windows are proxies of child processes (in
this case, cycling from a child-process window
should not deactivate that window if there is no
visible application window to activate).