Migrate PreferenceValue to ARC.
Enabled automatic ref. counting for PreferenceValue.
Also, modernized the headers (nullable annotations,
etc.), removing data members from the headers and
relocating documentation to property declarations.

Improved documentation for some of the bindings.

(This module is being ported for now but it is
becoming obsolete as more code migrates to SwiftUI.)