Make SWIG dependency more flexible; move to 4.0.2.
The build now checks two possible locations for
SWIG, where the default fallback is /usr/local.
In addition, the run script verifies that the
selected 'swig' is executable, and warns if it
does not match the last tested version (though
the build still allows any version to be used).

This mechanism has been tested with the latest
build installed by 'brew install swig' (4.0.2);
the project default SWIG version is now 4.0.2.