Add Notifications toolbar item, and MacTerm Help.
A new toolbar item for Notifications has been
added, as a short-cut for the existing commands
in the Terminal menu (such as “Notify on Next
Activity”).  As with other toolbar items, the
icon indicates what will happen when the button
is pressed.  Clicking will arbitrarily choose
“Notify on Next Activity”, at which point the
icon changes to a crossed-out bell; a menu on
the item can be used to select any other mode,
and clicking the crossed-out bell will turn off
any active mode.

Currently artwork is only displayed on macOS 11
(where system symbols are available); this may
display just a menu on macOS 10.15.

Unfortunately most of the convenient code to
handle session-dependent toolbar items had to
be replicated in another base class that uses
NSMenuToolbarItem, since the original was a
class inheriting from NSToolbarItem directly.

The Event Handler options were not previously in
MacTerm Help so a section has been added to
describe them all.