Improve gradient button states.
It is now possible to tie a keypad-style
button to a selected state, similar to
old-style square buttons.  (This is used
in places like key-binding preferences.)
Also improved the appearance of the
inactive state in light color schemes.
Added a new exact-width option to add
flexibility in using gradient buttons.