Migrate macro editor to SwiftUI.
The SwiftUI version of the editor looks nicer.
It also has new capabilities, such as dynamic
display of help based on selected action and
clearer annotation when the Control Keys panel
is bound.

The common mechanism for displaying option
lines in SwiftUI has been extended to allow
the “title” region to be a paragraph of small
help text instead.  This is used in the macro
editor but will also be useful when the
Translations pane is ported to SwiftUI later.

Fixed an issue with the preference for macro
key bindings in the case of an empty string,
which was triggering unnecessary errors and
logging.

Simplified macro editor panel’s header file.